+++ markdown_reader/preview.html
@@ -1,0 +1,184 @@
+
+            <html>
+            <head>
+                <meta charset="UTF-8">
+                <style>
+                    body {
+                        background-color: white;
+                        color: black;
+                        font-family: sans-serif;
+                        padding: 20px;
+                        font-size: 32px;
+                        line-height: 1.6;
+                    }
+                    pre, code {
+                        font-size: 28px;
+                        white-space: pre-wrap;
+                    }
+                    table {
+                        border-collapse: collapse;
+                        width: 100%;
+                        margin-top: 20px;
+                    }
+                    th, td {
+                        text-align: left;
+                        border: 1px solid #ccc;
+                        padding: 12px 16px;
+                        vertical-align: top;
+                    }
+                    th {
+                        background-color: #f3f3f3;
+                        color: #333;
+                    }
+                    tr:nth-child(even) {
+                        background-color: #fafafa;
+                    }
+                </style>
+            </head>
+            <body><p>当然可以，下面是对你提供的 PHP / Laravel 代码的<strong>中文逐行详细解释</strong>。</p>
+
+<p>这段代码定义了一个 <code>Controller</code> 基类，是<strong>所有控制器的父类</strong>，用于统一提供一些常用功能。</p>
+
+<hr />
+
+<h3>✅ 代码部分</h3>
+
+<div class="codehilite">
+<pre><span></span><code><span class="cp">&lt;?php</span>
+</code></pre>
+</div>
+
+<p>这是 PHP 文件的起始标记。</p>
+
+<hr />
+
+<div class="codehilite">
+<pre><span></span><code><span class="x">namespace App\Http\Controllers;</span>
+</code></pre>
+</div>
+
+<p>这是<strong>命名空间声明</strong>，说明该类属于 <code>App\Http\Controllers</code> 命名空间。在 Laravel 中，控制器都位于这个命名空间下，符合 PSR-4 自动加载规范。</p>
+
+<hr />
+
+<div class="codehilite">
+<pre><span></span><code><span class="x">use Illuminate\Foundation\Auth\Access\AuthorizesRequests;</span>
+<span class="x">use Illuminate\Foundation\Bus\DispatchesJobs;</span>
+<span class="x">use Illuminate\Foundation\Validation\ValidatesRequests;</span>
+<span class="x">use Illuminate\Routing\Controller as BaseController;</span>
+</code></pre>
+</div>
+
+<p>这几行是<strong>导入（use）Laravel 框架中的一些功能模块</strong>：</p>
+
+<table>
+<thead>
+<tr>
+  <th>导入模块</th>
+  <th>功能描述</th>
+</tr>
+</thead>
+<tbody>
+<tr>
+  <td><code>AuthorizesRequests</code></td>
+  <td>提供授权（权限）处理的方法，例如 <code>$this-&gt;authorize()</code>，用于控制用户是否有权限执行某些操作</td>
+</tr>
+<tr>
+  <td><code>DispatchesJobs</code></td>
+  <td>允许你调度 Laravel 的异步任务或队列（Jobs），如：<code>dispatch(new JobName)</code></td>
+</tr>
+<tr>
+  <td><code>ValidatesRequests</code></td>
+  <td>提供请求验证的方法，比如 <code>$this-&gt;validate($request, [...])</code></td>
+</tr>
+<tr>
+  <td><code>Controller as BaseController</code></td>
+  <td>引入 Laravel 框架的基础控制器类，并将其重命名为 <code>BaseController</code>，以避免命名冲突</td>
+</tr>
+</tbody>
+</table>
+
+<hr />
+
+<div class="codehilite">
+<pre><span></span><code><span class="x">class Controller extends BaseController</span>
+</code></pre>
+</div>
+
+<p>定义一个名为 <code>Controller</code> 的类，并<strong>继承</strong>自 <code>BaseController</code>（Laravel 框架的核心控制器）。</p>
+
+<p>这个类是整个应用中所有控制器的<strong>基类</strong>，意味着其他控制器（如 <code>HomeController</code>, <code>UserController</code>）通常会 <code>extends Controller</code> 来继承这些基础功能。</p>
+
+<hr />
+
+<div class="codehilite">
+<pre><span></span><code><span class="x">{</span>
+<span class="x">    use AuthorizesRequests, DispatchesJobs, ValidatesRequests;</span>
+<span class="x">}</span>
+</code></pre>
+</div>
+
+<p>这里使用 PHP 的 <code>trait</code> 特性（相当于模块引入），引入了上面导入的三个功能模块：</p>
+
+<table>
+<thead>
+<tr>
+  <th>trait 名称</th>
+  <th>功能作用</th>
+</tr>
+</thead>
+<tbody>
+<tr>
+  <td><code>AuthorizesRequests</code></td>
+  <td>可以在控制器中使用 <code>$this-&gt;authorize('update', $post)</code> 来检查用户权限</td>
+</tr>
+<tr>
+  <td><code>DispatchesJobs</code></td>
+  <td>允许控制器中派发任务到队列，例如 <code>dispatch(new SendEmailJob($user))</code></td>
+</tr>
+<tr>
+  <td><code>ValidatesRequests</code></td>
+  <td>提供请求验证的简洁写法 <code>$this-&gt;validate($request, ['name' =&gt; 'required'])</code></td>
+</tr>
+</tbody>
+</table>
+
+<hr />
+
+<h3>📌 总结</h3>
+
+<p>这段代码的作用是：</p>
+
+<ul>
+<li>定义一个 Laravel 中所有控制器的<strong>基类 Controller</strong></li>
+<li>它继承了 Laravel 框架的核心控制器类</li>
+<li>使用了 3 个非常实用的功能 trait（授权、任务调度、请求验证）</li>
+</ul>
+
+<p>其他控制器只要继承它，就自动具备这三个功能，非常方便开发。</p>
+
+<hr />
+
+<h3>✅ 举个例子：如果你写一个控制器</h3>
+
+<div class="codehilite">
+<pre><span></span><code><span class="x">class PostController extends Controller</span>
+<span class="x">{</span>
+<span class="x">    public function store(Request $request)</span>
+<span class="x">    {</span>
+<span class="x">        $this-&gt;validate($request, [&#39;title&#39; =&gt; &#39;required&#39;]); // 使用 ValidatesRequests 中的功能</span>
+<span class="x">        $this-&gt;authorize(&#39;create&#39;, Post::class); // 使用 AuthorizesRequests</span>
+<span class="x">        dispatch(new NotifyAdminJob($request-&gt;user())); // 使用 DispatchesJobs</span>
+<span class="x">    }</span>
+<span class="x">}</span>
+</code></pre>
+</div>
+
+<p>上面的控制器就可以直接使用父类提供的验证、授权和任务调度功能，得益于 <code>Controller</code> 类中的 <code>trait</code>。</p>
+
+<hr />
+
+<p>如你需要我展开解释某个 <code>trait</code> 的源码、示例用法，或者如何重写这个基类，也可以继续告诉我。</p>
+</body>
+            </html>
+
