+++ markdown_reader/utils.py
@@ -1,0 +1,13 @@
+import os
+import sys
+
+def get_resource_path(filename):
+    if hasattr(sys, '_MEIPASS'):
+        return os.path.join(sys._MEIPASS, filename)
+    elif getattr(sys, 'frozen', False):
+        return os.path.join(os.path.dirname(sys.executable), filename)
+    else:
+        return os.path.join(os.path.abspath("."), filename)
+
+def get_preview_file():
+    return os.path.join(os.path.dirname(os.path.abspath(__file__)), "preview.html")
